# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reconcile AI Apis",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/v1/reconcile": {
            "post": {
                "tags": [
                    "Reconciliation"
                ],
                "summary": "Reconcile two CSV or Excel files",
                "description": "Upload and compare two files using various reconciliation methods",
                "operationId": "4901856bb5a62aa674d08ea00562ea29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file1",
                                    "file2"
                                ],
                                "properties": {
                                    "file1": {
                                        "description": "First CSV or Excel file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file2": {
                                        "description": "Second CSV or Excel file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "reconcile_option": {
                                        "description": "Reconciliation method to use",
                                        "type": "string",
                                        "enum": [
                                            "reconcile_with_Gemini",
                                            "reconcile_with_recox_ai",
                                            "reconcile_with_openAI",
                                            "reconcile_with_deepSeek"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful reconciliation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reconciliation successful"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "matches": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "only_in_file1": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "only_in_file2": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reconciliations/{reconciliation}/export": {
            "post": {
                "tags": [
                    "Reconciliation"
                ],
                "summary": "Export reconciled data as a CSV file",
                "description": "Generates a CSV file containing matched and unmatched transactions from reconciliation.",
                "operationId": "07922d063e1c02d8f98dda6bf0ac25b4",
                "parameters": [
                    {
                        "name": "reconciliation",
                        "in": "path",
                        "description": "Reconciliation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file generated successfully",
                        "headers": {
                            "Content-Disposition": {
                                "description": "attachment; filename=reconciled-data.csv",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error while generating CSV file"
                    }
                }
            }
        },
        "/api/v1/reconciliations/{reconciliation}": {
            "get": {
                "tags": [
                    "Reconciliation"
                ],
                "summary": "Get reconciled records",
                "description": "Fetch reconciled records for the logged-in user",
                "operationId": "d158c4f3fdf513386913ea0159587c61",
                "parameters": [
                    {
                        "name": "reconciliation",
                        "in": "path",
                        "description": "Reconciliation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reconciled records fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reconciled records fetched successfully"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}